pyproject.toml
[tool.poetry]
name = "currency-exchange-fastapi"
version = "0.1.0"
description = ""
authors = ["Pavel Bramanis <bramanis2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
pydantic-settings = "^2.7.0"
black = "^24.10.0"
mypy = "^1.14.1"
python-multipart = "^0.0.20"
flake8 = "^7.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
